+++ CHANGELOG.md
@@ -6,6 +6,8 @@ This document describes the changes to Minimq between releases.
 
 ## Added
 * Support for the `Will` message specification.
+* [breaking] Adding `retained` flag to `publish()` to allow messages to be published in a retained
+  manner.
 
 # Version 0.4.0
 Version 0.4.0 was published on 2021-10-08